requirements.txt
Flask
gunicorn
streamlit
pandas
google-cloud-bigquery
google-cloud-aiplatform
google-auth
pytrends
beautifulsoup4
python-docx
db-dtypes
google-genai
google-cloud-storage
google-cloud-discoveryengine
google-api-core
pytz
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
cloud-sql-python-connector[pg8000]==1.6.0
pg8000==1.31.2